ci_cd_django/requirements.txt
asgiref==3.2.7
atomicwrites==1.4.0
attrs==19.3.0
colorama==0.4.3
Django==2.0.2
djangorestframework==3.11.0
gunicorn==20.0.4
importlib-metadata==1.6.0
more-itertools==8.3.0
packaging==20.4
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.2
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
wcwidth==0.2.2
zipp==3.1.0
